examples/gpio_monitor/requirements.txt
pkg-resources==0.0.0
RPi.GPIO==0.6.5
